+++ src/app/app.component.html
@@ -12,6 +12,3 @@
 -->
 <!-- Main Router -->
 <router-outlet></router-outlet>
-<ng-container *ngIf="isBrowser">
-  <p-confirmDialog [style]="{ width: '50vw' }"></p-confirmDialog>
-</ng-container>
