--- src/app/components/components.module.ts
@@ -1,37 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-import { VendorModule } from '../vendor.module';
-import { SharedModule } from '../shared/shared.module';
-
-// Global modals
-import { LogoutModalComponent } from './modals/logout/logout-modal.component';
-import { FeedbackModalComponent } from './modals/feedback/feedback.component';
-
-// Layout
-import { FooterComponent } from './masterpage/footer/footer.component';
-import { HeaderComponent } from './masterpage/header/header.component';
-import { LayoutMainComponent } from './masterpage/main/layout-main.component';
-import { LayoutSingleComponent } from './masterpage/single/layout-single.component';
-import { NavComponent } from './masterpage/nav/nav.component';
-
-// Modals include
-const APP_MODALS = [LogoutModalComponent, FeedbackModalComponent];
-
-// Components to include
-export const APP_COMPONENTS = [...APP_MODALS, FooterComponent, HeaderComponent, LayoutMainComponent, LayoutSingleComponent, NavComponent];
-
-@NgModule({
-    imports: [
-        // Angular
-        CommonModule,
-        // Shared
-        SharedModule,
-        // Vendors
-        VendorModule,
-    ],
-    providers: [],
-    declarations: [APP_COMPONENTS],
-    exports: [APP_COMPONENTS]
-})
-export class ComponentsModule {}
