+++ src/app/components/masterpage/index.ts
@@ -3,3 +3,4 @@ export * from './single/layout-single.component';
 export * from './header/header.component';
 export * from './nav/nav.component';
 export * from './footer/footer.component';
+export * from './masterpage.module';
