+++ src/app/components/masterpage/main/layout-main.component.html
@@ -1,3 +1,6 @@
 <app-header> <app-nav class="nav-component"></app-nav> </app-header>
 <main><router-outlet></router-outlet></main>
 <app-footer></app-footer>
+<ng-container *ngIf="isBrowser">
+  <p-confirmDialog [style]="{ width: '50vw' }"></p-confirmDialog>
+</ng-container>
