+++ src/app/components/masterpage/masterpage.routes.ts
@@ -1,0 +1,7 @@
+import { ModuleWithProviders } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+import { MasterPageModule } from './masterpage.module';
+
+const routes: Routes = [];
+
+export const routing: ModuleWithProviders<MasterPageModule> = RouterModule.forChild(routes);
