+++ src/app/shared/index.ts
@@ -1,0 +1,6 @@
+export * from './shared.module';
+export * from './directives';
+export * from './utils';
+export * from './stores';
+export * from './pipes';
+export * from './services';
