+++ tsconfig.json
@@ -20,13 +20,15 @@
     "lib": ["es2020", "dom"],
     "typeRoots": ["node_modules/@types"],
     "paths": {
+      "$components": ["./src/app/components"],
+      "$shared": ["./src/app/shared"],
+      "$env": ["./src/environments/environment.ts"],
       "$libs": ["./src/app/libs"],
       "$site": ["./src/app/site.module"],
       "$ui": ["./src/app/shared/stores/ui"],
       "$domain": ["./src/app/shared/stores/domain"],
       "$settings": ["./src/app/shared/stores/settings"],
       "$utils": ["./src/app/shared/utils"],
-      "$env": ["./src/environments/environment.ts"],
       "$models": ["./src/shared/models"],
       "$modals": ["./src/compoments/modals"]
     }
